Refactor workflow
